python/requirements.txt
pyzmq>=24.0.0
numpy>=1.21.0
pyqtgraph>=0.13.0
PySide6>=6.4.0